+++ src/plugins/odklad/4-pokracovat/pokracovat.tsx
@@ -1,0 +1,83 @@
+import React from 'react'
+import getConfig from 'next/config'
+import { Page } from '../../../components/Page'
+import { PostponeUserInput } from '../../../types/PostponeUserInput'
+import { convertPostponeToXML } from '../../../lib/postpone/postponeConverter'
+import { RedirectField, RedirectForm } from '../../_components/form/RedirectForm'
+import { setDate, toBase64 } from '../../../lib/utils'
+import { TAX_YEAR } from '../../../lib/calculation'
+
+const {
+  publicRuntimeConfig: { odkladEmailTemplateId },
+} = getConfig()
+
+const buildXml = (userInput) => convertPostponeToXML(setDate(userInput))
+
+const buildFields = (userInput: PostponeUserInput): RedirectField[] => {
+  const CALLBACK_PATH =
+    '/zivotne-situacie/odklad-danoveho-priznania/krok/registrovat-sa-na-financnej-sprave'
+
+  const fullName = `${userInput.meno} ${userInput.priezvisko}`
+  const xmlFile = toBase64(buildXml(userInput))
+  const deadline = userInput.prijmy_zo_zahranicia
+    ? `30. september ${TAX_YEAR + 1}`
+    : `30. jún ${TAX_YEAR + 1}`
+
+  return [
+    { name: 'submission[type]', value: 'EmailMeSubmissionInstructionsEmail' },
+    {
+      name: 'submission[callback_url]',
+      value: CALLBACK_PATH,
+    },
+    {
+      name: 'submission[callback_step_path]',
+      value: CALLBACK_PATH,
+    },
+    {
+      name: 'submission[callback_step_status]',
+      value: 'not_started',
+    },
+    {
+      name: 'submission[attachments[]filename]',
+      value: 'odklad-danoveho-priznania.xml',
+    },
+    { name: 'submission[attachments[]body_base64]', value: xmlFile },
+    {
+      name: 'submission[subscription_types][]',
+      value: 'EmailMeSubmissionInstructionsEmail',
+    },
+    {
+      name: 'submission[subscription_types][]',
+      value: 'SelfEmployedSubscription',
+    },
+    {
+      name: 'submission[subscription_types][]',
+      value: 'NewsletterSubscription',
+    },
+    {
+      name: 'submission[extra][template_id]',
+      value: odkladEmailTemplateId,
+    },
+    {
+      name: 'submission[extra][params][recipient_name]',
+      value: fullName,
+    },
+    {
+      name: 'submission[extra][params][deadline]',
+      value: deadline,
+    },
+  ]
+}
+
+const ContinuePage: Page<{}> = ({ postponeUserInput, isDebug }) => {
+  return (
+    <RedirectForm
+      isDebug={isDebug}
+      fields={buildFields(postponeUserInput)}
+      canContinue={!!postponeUserInput.priezvisko}
+      debugDownload={buildXml(postponeUserInput)}
+    />
+  )
+}
+
+export { ContinuePage }
